requirements.txt
absl-py==0.13.0
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.4
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.2
importlib-metadata==4.8.1
isort==5.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas
Pillow==8.3.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6
tensorflow-estimator==2.6.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.5.0
gunicorn==19.9.0
gdown